Centered cards
i used flexbox to center all the cards. for some reason i found that the cards on the front page where not center, i dont know if it was like this before i started changing the position of the cards on the user page. Anyway i didnt do anything to the css for the CardContainer on the front page. I also changed it so that its no description on the cards, just country and city